Refactored project structure to match onion architecture
