+++ package.json
@@ -7,12 +7,12 @@
     "test": "test"
   },
   "scripts": {
-    "start": "fastify start -l info dist/app.js",
-    "start:ts": "npm run build && fastify start -l info dist/app.js",
+    "start": "fastify start -l info dist/presentation/http/app.js",
+    "start:ts": "npm run build && fastify start -l info dist/presentation/http/app.js",
     "build": "tsc",
     "watch:ts": "tsc -w",
     "dev": "npm run build && concurrently -k -p \"[{name}]\" -n \"TypeScript,App\" -c \"yellow.bold,cyan.bold\" \"npm:watch:ts\" \"npm:dev:start\"",
-    "dev:start": "fastify start --ignore-watch=.ts$ -w -l info -P dist/app.js",
+    "dev:start": "fastify start --ignore-watch=.ts$ -w -l info -P dist/presentation/http/app.js",
     "lint": "prettier --write {src,test}/**/*.{ts,js,json} && eslint {src,test}/**/*.{ts,js,json}",
     "test:unit": "jest test/unit --verbose --coverage",
     "test:integration": "jest test/integration --verbose --coverage",
