+++ src/application/utils/staticInterfaces.ts
