+++ src/domain/models/rate.ts
@@ -1,0 +1,3 @@
+type rate = number | null;
+
+export { rate };
