--- src/interfaces/currencyRateService.ts
@@ -1,28 +1,0 @@
-type rate = number | null;
-
-interface ICurrencyRateServiceFactory {
-  createCurrencyRateService(
-    api_key: string
-  ): ICurrencyRateService & IChainedCurrencyRateService;
-}
-
-interface ICurrencyRateService {
-  getRate(): Promise<rate>;
-  getProviderInfo(): string;
-}
-
-interface IHTTPCurrencyRateService {
-  readonly url: string;
-}
-
-interface IChainedCurrencyRateService {
-  next: (IChainedCurrencyRateService & ICurrencyRateService) | null;
-}
-
-export {
-  ICurrencyRateServiceFactory,
-  ICurrencyRateService,
-  IHTTPCurrencyRateService,
-  IChainedCurrencyRateService,
-  rate
-};
