--- src/interfaces/subscriptionRepository.ts
@@ -1,11 +1,0 @@
-import { ISubscribtionModel } from "./SubscribtionModel";
-
-interface ISubscribtionRepository {
-  Model: unknown;
-  findAll(): Promise<Array<ISubscribtionModel>>;
-  findOneByEmail(email: string): Promise<ISubscribtionModel | undefined>;
-  createOne(email: string): Promise<ISubscribtionModel | undefined>;
-  serializeToEmails(subscribtions: Array<ISubscribtionModel>): Array<string>;
-}
-
-export { ISubscribtionRepository };
