--- src/repositories/index.ts
@@ -1,6 +1,0 @@
-import SubscribtionModelFS from "../database/models/subscribtion.model";
-import { SubscribtionRepository } from "./subscribtion.repository";
-
-const subscribtionRepository = new SubscribtionRepository(SubscribtionModelFS);
-
-export { subscribtionRepository };
