--- src/routes/sendEmails/index.ts
@@ -1,23 +1,0 @@
-import { FastifyPluginAsync } from "fastify";
-import { mainCurrencyRateService } from "../../services/currecyRate.service";
-import { emailsSenderService } from "../../services/emailSender.service/index";
-import { HttpResponseMessage } from "../../utils/httpResponseMessage.enum";
-import { subscribtionRepository } from "../../repositories";
-
-const sendRateEmails: FastifyPluginAsync = async (fastify): Promise<void> => {
-  fastify.post("/", sendRateEmailsHandler);
-};
-
-export default sendRateEmails;
-
-async function sendRateEmailsHandler() {
-  const rate = await mainCurrencyRateService.getRate();
-  const subject = "Exchange rate: BTC to UAH";
-  const mailingListOptions = { subject, html: rate?.toString() };
-
-  const subscribtions = await subscribtionRepository.findAll();
-  const emails = subscribtionRepository.serializeToEmails(subscribtions);
-
-  await emailsSenderService.sendMailingList(mailingListOptions, emails);
-  return { status: "success", message: HttpResponseMessage.EMAILS_SENT };
-}
